Actualización 18/07/2024 - Añadido relación roles - modificar.html. Ahora cuando entramos a crear usuario en Admin se puede establecer una relación entre usuario y rol.

+++ src/test/java/com/eoi/Fruggy/EntidadesTest/RolTest.java
@@ -35,7 +35,7 @@ public class RolTest {
 
         repoRol.save(guardarRol);
 
-        Rol rolguardado = repoRol.findById((int) guardarRol.getId()).get();
+        Rol rolguardado = repoRol.findById((long) guardarRol.getId()).get();
 
         assertEquals("Admin", rolguardado.getRolNombre());
         assertNotNull(rolguardado.getUsuariosRol());
